cleaned up how dbs are deleted.

+++ Scripts/Menu/menu-applications-manage.sh
@@ -128,9 +128,8 @@ EOF
                     _sleep
 
                     bash -c "$HOME"/RoninDojo/Scripts/Install/install-electrs-indexer.sh
-                    sudo rm -rf "${docker_volume_indexer}"/_data/addrindexrs
-                    sudo rm -rf "${docker_volume_indexer}"/_data/electrs
-                    sudo rm -rf "${docker_volume_indexer}"/_data/db/mainnet
+                    test -d "${docker_volume_indexer}"/_data/db/mainnet && sudo rm -rf "${docker_volume_indexer}"/_data/db/mainnet
+                    test -d "${docker_volume_indexer}"/_data/addrindexrs && sudo rm -rf "${docker_volume_indexer}"/_data/addrindexrs
                     ;;
                 "Install Indexer")
                     cat <<EOF
